# api/openapi.yaml
openapi: 3.1.0
info:
  title: Codex CPA Pool v2 API
  version: 2.0.0-dev
  description: >-
    Contract for the Go Admin, self-service and public configuration APIs.
    Authored response text uses Accept-Language (en or zh-CN), defaults to en, and
    declares Content-Language and Vary: Accept-Language. Codes, IDs and stored values
    are language independent. Clients must not parse display messages for control flow.
servers:
  - url: /
...
paths:
...
  /admin/api/accounts/model-test:
    post:
      operationId: testAccountModel
      tags: [admin-accounts]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountModelTestRequest"
      responses:
        "200":
          description: Fixed short generation test, bounded to 25 seconds; upstream failures are returned as unsuccessful results without invalidating the Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountModelTestResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: cpa_admin_session
    portalSession:
      type: apiKey
      in: cookie
      name: cpa_usage_session
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
...
  responses:
...
    ErrorResponse:
      description: Safe JSON error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message_key: { type: string, description: Stable authored message identifier }
        message_params: { type: object, additionalProperties: true }
        message:
          type: string
        type:
          type: string
...
    AccountModelTestRequest:
      type: object
      required: [account, model]
      properties:
        account: { type: string }
        model: { type: string, minLength: 1, maxLength: 128 }
    AccountModelTestResult:
      type: object
      required: [account, model, success, elapsed_ms, checked_at, upstream_status, code, message]
      properties:
        account: { type: string }
        model: { type: string }
        success: { type: boolean }
        elapsed_ms: { type: integer, format: int64 }
        checked_at: { type: integer, format: int64 }
        upstream_status: { type: integer }
        code: { type: string }
        message: { type: string }
        message_key: { type: string, description: Stable authored message identifier }
        message_params: { type: object, additionalProperties: true }
...